adds item analyst classes
